refactor: 패키지 구조 개편 devjun10/issue-tracker#11
