--- Backend/src/main/java/com/team33/backend/domain/Milestone.java
@@ -1,39 +1,0 @@
-package com.team33.backend.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.validation.constraints.NotBlank;
-import java.time.LocalDate;
-import java.util.Objects;
-
-@Entity
-public class Milestone extends CommonEntity {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length = 255)
-    @NotBlank
-    private String title;
-
-    @Column(length = 8192)
-    private String description;
-
-    private LocalDate deadLine;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Milestone)) return false;
-        Milestone entity = (Milestone) o;
-        return Objects.equals(id, entity.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
