+++ Backend/src/main/java/com/team33/backend/emoji/Emoji.java
@@ -1,5 +1,9 @@
-package com.team33.backend.domain;
+package com.team33.backend.emoji;
 
+import com.team33.backend.comment.Comment;
+import com.team33.backend.issue.domain.Issue;
+
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
@@ -10,25 +14,28 @@ import javax.persistence.ManyToOne;
 import java.util.Objects;
 
 @Entity
-public class IssueGroupMember {
+public class Emoji {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "group_id", nullable = false)
-    private IssueGroup issueGroup;
+    @Column(length = 4)
+    private String emotion;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
+    @JoinColumn(name = "issue_id", nullable = false)
+    private Issue issue;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "comment_id")
+    private Comment comment;
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (!(o instanceof IssueGroupMember)) return false;
-        IssueGroupMember entity = (IssueGroupMember) o;
+        if (!(o instanceof Emoji)) return false;
+        Emoji entity = (Emoji) o;
         return Objects.equals(id, entity.id);
     }
