+++ Backend/src/main/java/com/team33/backend/issue/domain/Milestone.java
@@ -1,37 +1,36 @@
-package com.team33.backend.domain;
+package com.team33.backend.issue.domain;
+
+import com.team33.backend.common.CommonEntity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
 import javax.validation.constraints.NotBlank;
-import java.util.ArrayList;
-import java.util.List;
+import java.time.LocalDate;
 import java.util.Objects;
 
 @Entity
-public class Member {
+public class Milestone extends CommonEntity {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(length = 39)
+    @Column(length = 255)
     @NotBlank
-    private String name;
+    private String title;
 
-    @Column(length = 2048)
-    private String profileImageUrl;
+    @Column(length = 8192)
+    private String description;
 
-    @OneToMany(mappedBy = "member")
-    private List<Issue> issues = new ArrayList<>();
+    private LocalDate deadLine;
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (!(o instanceof Member)) return false;
-        Member entity = (Member) o;
+        if (!(o instanceof Milestone)) return false;
+        Milestone entity = (Milestone) o;
         return Objects.equals(id, entity.id);
     }
