+++ Backend/src/main/java/com/team33/backend/member/Member.java
@@ -1,35 +1,39 @@
-package com.team33.backend.domain;
+package com.team33.backend.member;
+
+import com.team33.backend.issue.domain.Issue;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.OneToMany;
 import javax.validation.constraints.NotBlank;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 
 @Entity
-public class Label {
+public class Member {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(length = 50)
+    @Column(length = 39)
     @NotBlank
     private String name;
 
-    @Column(length = 100)
-    private String description;
+    @Column(length = 2048)
+    private String profileImageUrl;
 
-    @Column(length = 6, columnDefinition = "char(6)")
-    @NotBlank
-    private String color;
+    @OneToMany(mappedBy = "member")
+    private List<Issue> issues = new ArrayList<>();
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (!(o instanceof Label)) return false;
-        Label entity = (Label) o;
+        if (!(o instanceof Member)) return false;
+        Member entity = (Member) o;
         return Objects.equals(id, entity.id);
     }
